Add vue-router with /test page for plugin testing
Introduce vue-router to enable multiple pages. Move main app content
to HomeView.vue and add TestView.vue at /test for testing plugins
individually with a select dropdown and JSON argument input.
