+++ .gitignore
@@ -5,3 +5,4 @@ build
 .DS_Store
 output
 benchmark/results
+*~
