+++ src/main.ts
@@ -1,6 +1,7 @@
 import { createApp } from "vue";
 import App from "./App.vue";
+import router from "./router";
 import "./index.css";
 import "material-icons/iconfont/material-icons.css";
 
-createApp(App).mount("#app");
+createApp(App).use(router).mount("#app");
