+++ src/router/index.ts
@@ -1,0 +1,21 @@
+import { createRouter, createWebHistory } from "vue-router";
+import HomeView from "../views/HomeView.vue";
+import TestView from "../views/TestView.vue";
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes: [
+    {
+      path: "/",
+      name: "home",
+      component: HomeView,
+    },
+    {
+      path: "/test",
+      name: "test",
+      component: TestView,
+    },
+  ],
+});
+
+export default router;
